Update project description
Started on the project description. May have more info than necessary (or irrelevant info) but could be good for us to look at it and say what is good and bad, or if there are any misconceptions about the project that need to be cleared up. Also copy-pasted the executive summary in here (which needs changes, and we probably won't use most of it) but is pretty much the technical/business problem that needs to be solved and the benefits.